requirements.txt
Flask~=1.1.2
Flask-Cors~=3.0.10
zappa~=0.52.0
boto3~=1.17.22
PyGithub~=1.54.1
